Added FTS on book title and author
I want a FTS solution that a) doesn't require storing a whole text file as a
string in database and/or in memory, and b) doesn't require a separate
service like Solr or Elsticsearch. Unfortunately there's no equivalent of
Lucene in JavaScript -- there are a few packges that index JSON but not
files. So for this implementation I just use the MongoDB built-in text search
on the title and the author field. The actual file is not searched, and
Chinese is not supported (according to MongoDB documentation, text search
in Chinese is supported in the Enterprise version).